requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-huggingface
langchain-deepseek

sentence-transformers
transformers
accelerate
huggingface-hub

psycopg2-binary
pgvector

pydantic
pypdf